Update logic to disconnect URLs from Batches
