+++ src/api/endpoints/batch/routes.py
@@ -4,12 +4,12 @@ from fastapi import Path, APIRouter
 from fastapi.params import Query, Depends
 
 from src.api.dependencies import get_async_core
-from src.api.endpoints.batch.dtos.get.duplicates import GetDuplicatesByBatchResponse
 from src.api.endpoints.batch.dtos.get.logs import GetBatchLogsResponse
 from src.api.endpoints.batch.dtos.get.summaries.response import GetBatchSummariesResponse
 from src.api.endpoints.batch.dtos.get.summaries.summary import BatchSummary
-from src.api.endpoints.batch.dtos.get.urls import GetURLsByBatchResponse
 from src.api.endpoints.batch.dtos.post.abort import MessageResponse
+from src.api.endpoints.batch.duplicates.dto import GetDuplicatesByBatchResponse
+from src.api.endpoints.batch.urls.dto import GetURLsByBatchResponse
 from src.collectors.enums import CollectorType
 from src.core.core import AsyncCore
 from src.core.enums import BatchStatus
