+++ src/api/endpoints/metrics/batches/aggregated/dto.py
