+++ src/api/endpoints/metrics/batches/breakdown/dto.py
