+++ src/db/client/helpers.py
@@ -1,0 +1,3 @@
+def add_standard_limit_and_offset(statement, page, limit=100):
+    offset = (page - 1) * limit
+    return statement.limit(limit).offset(offset)
