--- tests/automated/integration/api/test_duplicates.py
@@ -1,63 +1,0 @@
-import pytest
-
-from src.api.endpoints.batch.dtos.get.summaries.summary import BatchSummary
-from src.db.dtos.batch import BatchInfo
-from src.collectors.source_collectors.example.dtos.input import ExampleInputDTO
-from tests.automated.integration.api.conftest import disable_task_trigger
-
-
-@pytest.mark.asyncio
-async def test_duplicates(api_test_helper):
-    ath = api_test_helper
-
-    # Temporarily disable task trigger
-    disable_task_trigger(ath)
-
-    dto = ExampleInputDTO(
-            sleep_time=0
-        )
-
-    batch_id_1 = ath.request_validator.example_collector(
-        dto=dto
-    )["batch_id"]
-
-    assert batch_id_1 is not None
-
-    batch_id_2 = ath.request_validator.example_collector(
-        dto=dto
-    )["batch_id"]
-
-    assert batch_id_2 is not None
-
-    await ath.wait_for_all_batches_to_complete()
-
-
-    bi_1: BatchSummary = ath.request_validator.get_batch_info(batch_id_1)
-    bi_2: BatchSummary = ath.request_validator.get_batch_info(batch_id_2)
-
-    bi_1.url_counts.total = 2
-    bi_2.url_counts.total = 0
-    bi_1.url_counts.duplicate = 0
-    bi_2.url_counts.duplicate = 2
-
-    url_info_1 = ath.request_validator.get_batch_urls(batch_id_1)
-    url_info_2 = ath.request_validator.get_batch_urls(batch_id_2)
-
-    assert len(url_info_1.urls) == 2
-    assert len(url_info_2.urls) == 0
-
-    dup_info_1 = ath.request_validator.get_batch_url_duplicates(batch_id_1)
-    dup_info_2 = ath.request_validator.get_batch_url_duplicates(batch_id_2)
-
-    assert len(dup_info_1.duplicates) == 0
-    assert len(dup_info_2.duplicates) == 2
-
-    assert dup_info_2.duplicates[0].original_url_id == url_info_1.urls[0].id
-    assert dup_info_2.duplicates[1].original_url_id == url_info_1.urls[1].id
-
-    assert dup_info_2.duplicates[0].original_batch_id == batch_id_1
-    assert dup_info_2.duplicates[1].original_batch_id == batch_id_1
-
-
-
-
