+++ tests/automated/integration/tasks/asserts.py
@@ -1,3 +1,4 @@
+from src.core.tasks.base.run_info import TaskOperatorRunInfo
 from src.core.tasks.url.enums import TaskOperatorOutcome
 
 
@@ -10,6 +11,6 @@ async def assert_prereqs_met(operator):
     assert meets_prereqs
 
 
-def assert_task_has_expected_run_info(run_info, url_ids: list[int]):
-    assert run_info.outcome == TaskOperatorOutcome.SUCCESS
+def assert_task_has_expected_run_info(run_info: TaskOperatorRunInfo, url_ids: list[int]):
+    assert run_info.outcome == TaskOperatorOutcome.SUCCESS, run_info.message
     assert run_info.linked_url_ids == url_ids
